# Conflicts:
#	Modules/TArc/Sources/TArcCore/Private/TArcCore.cpp
#	Sources/Internal/Scene3D/Scene.h


Former-commit-id: e497bec241fcb57154fab47fa2bb846c605cbe48

+++ Sources/External/cef/Sources/CEFController.cpp
@@ -195,9 +195,9 @@ void CEFControllerImpl::CleanCache()
     size_t cacheSize = 0;
     for (const FilePath& path : cacheDirContent)
     {
-        uint32 size = 0;
+        uint64 size = 0;
         fs->GetFileSize(path, size);
-        cacheSize += size;
+        cacheSize += static_cast<size_t>(size);
 
         if (cacheSize > cacheSizeLimit)
         {
